:bug: Drop UTC offset for local time used in tasks table (#1994)
The API provides date/time in UTC. For user convenience we display it in
the browser time zone. Before this PR, UTC offset for the local time
zone was included.

The format used is a subset of ISO 8601 which allows for predictable
sorting in the table.
Format string: "YYYY-MM-DD HH:mm:ss".
Example date: "2024-07-03 13:21:44".

Part-of: #1969

Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>