:seedling: Add data attributes to table rows (#1996)
To enable easier and safer selection of table rows by testing code, add
some data attributes to the tr:

  - `data-item-id` - Holds the `idProperty` value of a table's items. This
    is typically the item's id number but may be a frontend generated ID in
    some cases.

  - `data-item-name` - If the table property `dataNameProperty` is
    provided, hold the name property of a table's item.

For any tables that use `/use.?*TableControls\(/`, add a
`dataNameProperty` and if necessary, adjust the `idProperty`. A few
tables for items without a good item id and name property were left with
just the `idProperty`.

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>