:sparkles: Add handling for tasks "Succeeded with errors" in task and application tables (#2011)
### Summary
Resolves: #1997

When a task is succeeded but has attached errors, render task state,
application tasks state and application analysis state with a warning.

### Details of Changes
With this change, hub is not adding a new terminal state, but having a
new terminal state is an easy way to handle things. The task queries
will now evaluate the `state` and `errors` props to determine if
"Succeeded" should be "SucceededWithErrors". This synthetic task status
makes updating all of the UI component much simpler.

To function properly, HUB [Issue
725](https://github.com/konveyor/tackle2-hub/issues/725) / [PR
720](https://github.com/konveyor/tackle2-hub/pull/720) are required.

Changes have also been made to utilize the `/tasks/reports/dashboard`
endpoint to get a much smaller view of tasks when dealing with them in
aggregate.

The application details drawer was refactored to push data fetching
closer to actual use, and to break up a very large component into a
container component with a component dedicated to each tab.

### Screenshots
Application inventory page:
![screenshot-localhost_9000-2024 07
15-18_33_20](https://github.com/user-attachments/assets/b7015ebf-4305-4d6d-a670-5e4fe532981d)

Application tasks popover:
![image](https://github.com/user-attachments/assets/14d4d19e-eaca-4505-88be-d5016da1e70b)

Task Manager page:
![screenshot-localhost_9000-2024 07
15-18_35_15](https://github.com/user-attachments/assets/2547a46a-1fa6-456f-836c-4e59cc0fd074)

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>