:bug: Address targets falling out of sync after label selection changes (#2019)
Resolves: https://issues.redhat.com/browse/MTA-3180

Currently, the set-options screen allows edit of target labels which
can be set manually on the Advanced options step or by way of selecting
a target card on the target card step. This PR brings the form labels
and selectedTarget form values in sync by updating the selected targets
to reflect any form label changes in the set options page. This fixes
the mta-3180 scenario by causing validation to reflect the now accurate
state of the selectedTargets form value.

Signed-off-by: Ian Bolton <ibolton@redhat.com>