:bug: Archetype table, applications column is now a link (#2024)
When an archetype matches at least 1 application, render a link in the
application columns that will navigate the user to the application table
with a filter applied to show all of the applications matched to the
archetype.

The `Link` component is now shared between the archetype table and the
archetype detail drawer.

Resolves: https://issues.redhat.com/browse/MTA-1390

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>