:bug: Add application risk filter of "Unassessed"
Resolves: https://issues.redhat.com/browse/MTA-2816

Application risk options are one of:
  - High (red)
  - Medium (yellow)
  - Low (green)
  - Unknown (unknown)
  - Unassessed (unassessed)

Added the __Unassessed__ option to the risk filter
list on the application table to have a complete
set of risk filters.

Note: To help test risk levels, the questionnaire
`hack/import-questionnaire/assign-risk.yaml` has
been added.  The single question has a 1 to 1
mapping between answer and risk level.  As long as
that questionnaire is the only one active on the
system, it works well to quickly assign risks to
applications.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>