:seedling: Refactor dist npm script
Change the copy package used to generate the `dist/`
folder after building to use `copy-files-from-to`.

The package change allows us to have a configuration
file with multiple includes and excludes to make sure
we copy only exactly what is needed when the `dist/`
folder is built.  The contents of the `dist/` folder
is what is copied into the runtime container.  It
should be as light as possible.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>