:seedling: Skip PR CI actions for docs/hacks changes
When a PR only contains changes to root level markdown,
or anything under `docs/` or `hack/`, there is no need
to run the CI github actions.

To satisfy branch protection rules that may exist that
require the "unit-test" job to run, the `ci-repo.yml`
action looks up details about a PR and will skip the
unit test if it detects that the only changes made in
the PR are docs or hacks.

The ignores do not apply to pushes to main, pushes to release,
or workflow calls.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>