:seedling: [backport release-0.6] Update ci-global.yml (#2179)(#2186) (#2189)
Backport of: #2179
Backport of: #2186

With `upload-artifact@v3` deprecated[^1] and shut off, move to use the
`global-ci-bundle.yml` workflow for global CI. This change drops the
need for artifacts and just uses the ttl.sh service[^2]. This technique
is already being used in a few other Konveyor projects as well.

A recent change in the UI tests repo[^3] changed the basic CI testing
suite from `@tier0` to `@ci`. Updating the CI workflow to use the new
tier.

[^1]: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/
[^2]: https://ttl.sh
[^3]: https://github.com/konveyor/tackle-ui-tests/pull/1351

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>