:bug: [backport release-0.6] Update application assess override check and assessment column (#2205) (#2207)
Backport-of: #2205
Fixes: https://issues.redhat.com/browse/MTA-1825

__Summary__
On the application inventory page, update handling for the:
  - assessment column
  - assess an application action

__Assessment column__
Collect the information about an application's assessment status as part
of the `DecoratedApplication`. This allows easy access to the direct and
inherited assessments for an application. The Assessment column
component now uses this data for its rendering.

__Assess and application action__
When invoking the "Assess" action on an application, check the
assessment status:
- If the application is already being directly assessed or does not
belong to an archetype, push the user to the assessments page.
- If the application belongs to an archetype, and if none of the
archetypes have an assessment, push the user to the assessments page.
- If the application's archetypes have any assessments, prompt the user
if they want to override the archetype assessments or view the
archetype's assessments.

The `useAssessmentStatus()` hook was broken up with the logic to
`client/src/app/utils/application-assessment-status.ts` and the runtime
to `useDecoratedApplication()`.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>