:sparkles: Refactor uploading and validation of custom rules files
When uploading custom rule files as YAML or XML, make
sure that the file can be parsed as a YAML or XML file
before the file is accepted.  If a file cannot be validated,
display an error and flag the file as in error.  These files
will not pass to further processing.  Upload forms will
not be able to be submitted with an invalid file.

Supporting changes:
  - Add `CustomRuleFilesUpload` component to share rule
    file upload behavior in multiple places

  - Refactor `useRuleFiles()` hook to work well with
    `CustomRuleFilesUpload` and not keep internal state
    beyond what is necessary

  - api/rest functions: extract file related api calls to rest/files

------

Drop support for windup/xml useRuleFiles

  - Drop xmllint and support for it everywhere
  - Remove the XML file validation functions
  - Remove the XML file rule parsing
  - Drop xml as an accepted upload file on the custom rules files upload component

Now all of the xml validation/parsing is only in the
`maven-settings` utilities!

------

refactor the file upload mutations to take a File instead of FormData, other refactoring that makes sense

------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>