:bug: Validate shape of uploaded custom rules files
Resolves: https://issues.redhat.com/browse/MTA-4749

When uploading custom rule files aw YAML or XML, make
sure that the file can be read as a YAML or XML file
before the file is accepted.  If a file cannot be validated,
display an error and flag the file as in error.  These files
will not pass to further processing.  Upload forms will
not be able to be submitted with an invalid file.

Supporting changes:
  - Add `CustomRuleFilesUpload` component to share rule
    file upload behavior in multiple places

  - Refactor `useRuleFiles()` hook to work well with
    `CustomRuleFilesUpload` and not keep internal state
    beyond what is necessary

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>