:bug: when active tasks go inactive, invalidate the application query
Resolves: https://issues.redhat.com/browse/MTA-5181

For performance reasons, the application inventory page
does not always refresh in the background.  When there
are no active tasks or imports, the application query
does not auto-refresh.  At the point where the app goes
from having active tasks to no active tasks (e.g. running
an analysis), it is possible that the auto-refresh is
disabled before updates can be fetched.  This is at most
a 5 second window.

To avoid this, when the tasks dashboard query goes from having
active tasks to no active tasks, the application query is
invalidated forcing the query to refresh its data.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>