:bug: always refresh the application query on application inventory (#2346)
Resolves: https://issues.redhat.com/browse/MTA-5181

Historically and for performance reasons, the application inventory page
tried to not always refresh in the background. The approach never really
consistently worked, and if hub has a large number of applications, the
refresh query logic would almost always run anyway.

The refresh logic has been removed and the page will refresh the
applications every 5 seconds.

Performance issues may emerge in a large environment and #2350 was
written to capture what is known and have some ideas about how to
approach a solution.

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Signed-off-by: Cherry Picker <noreply@github.com>