:seedling: Update use of MSW in jest tests (#2360)
MSW provides responses to axios calls made in the various react queries
used by components being tested in our jest tests. Global handling of
the MSW handlers is setup in `setupTests.ts`. After each test, the MSW
server handlers are globally reset. Any handlers that are needed need in
a test suite either be setup in the test/it block directly, or globally
for a suite using `beforeEach()` in the suite's `define()` block.

Tests that use components with queries/axios calls have been updated to
this style. No axios network exception error should be logged to
console.

In future, if any axios network exceptions are logged to console, the
test should be updated to provide an msw handler as appropriate for the
test.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>