:bug: Fix handling of source platform coordinates on application form
Resolves: https://issues.redhat.com/browse/MTA-6100
Resolves: #2563

Change the source platform data flow such that the source platform
coordinates schemas are available prior to the selection of the
source platform.  This way, there is no delay in applying the initial
schema or changing the schema on selection/clearing of the source
platform.

Now form validation works as expected and when a source platform is
selected, and coordinates are entered, changes in other fields will
enable the save button.

Related changes:
- Add `useRepositoryKind` hook for reuse

  Even though the hook only holds static data, keeping
  it as a hook allows for reuse in other components.

- Update a lot of queries to report isLoading and isSuccess instead of isFetching

  isLoading, isSuccess, and isError are all data related.

  The isFetching status is based on when the fetch is actually running
  and may not result in any updates.  We really care more about having
  data than the fetch status.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>