:seedling: Add devtools/schema-defined page to test schema defined fields
In support of work on #2550, when `DEVTOOLS=on` env var is set,
create a devtools perspective with a "SDF Testing" page.  This
page allows for testing the `SchemaDefinedField` component against
a variety of JSON schemas and documents.  This should allow for
easier development and debugging of the `SchemaDefinedField` component.

Changes:
  - Add "devtools/schema-defined" page to work with schema defined fields
    in a tighter loop
  - Hide the tools behind the `DEVTOOLS` env var
  - Include a variety of example schemas
  - Add an index.ts file to export components/utils from schema-defined-fields
  - Update imports in relevant fields

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>