:bug: Add discover application wizard to the applications table
Resolves: https://issues.redhat.com/browse/MTA-6120
Resolves: #2629

The source platform discovery applications wizard is now
available in the applications table. A new step to select the
source platform has been added to the wizard. The wizard
itself is now a shared component that is both from the
applications table and the source platforms page.

Summary of changes:
  - Move the discover import wizard to a shared component
  - Add a reducer value/setter for the platform
  - Allow easy initialization of the wizard state
  - Add a select platform step to the wizard if no platform is provided
  - Add `usePlatformKindList` hook
  - Update `useRepositoryKind` hook to match i18n use

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>