:bug: Fix application source and asset repository validation
Resolves: https://issues.redhat.com/browse/MTA-6114
Resolves: https://issues.redhat.com/browse/MTA-6105
Resolves: https://issues.redhat.com/browse/MTA-5769
Resolves: #2615
Partially addresses: #2541

To properly support the input and validation of the 4 fields
needed for a source repository, the "repository-fields" components,
schemas, and tests were created.  Now, the source repository fields
and the asset repository fields are handled with exactly the same
schemas and components.

Now:
  - The type is defaulted to 'git' url validation
  - The type is required once the url, branch or path is entered
  - The url is required one the branch or path is entered
  - If the branch or path are empty, the url is allowed to be empty
  - If only the type is selected, the repository is considered empty

With the new Repository based schema, it can be shared anywhere
the 4 repo fields are needed.  A collection of unit tests cover the
main functionality of the schema.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>