:seedling: Add inspect_pods.sh hack script (#2716)
Add a script to inspect the pods and their container images.

Targets a Konveyor operator deployed to minikube. This script will not
work for OpenShift.

Inspects the pods and their container images. Check the remote container
repository for the image's manifest digest for the pod's image tag.
Report if they match or not.

A mismatch indicates that the operator needs to be updated.

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>