:bug: Update Cloud Foundry discover applications filter
Resolves: https://issues.redhat.com/browse/MTA-6326
Resolves: #2709
CloudFoundry schema update: https://github.com/konveyor/operator/pull/484

Update the discover application wizard CloudFoundry source platform
filter input and review to support organizations, spaces, and names.

Added the ability to validate that two JSON schemas are functionally
equivalent.  This is used to validate that the Cloud Foundry schema
is functionally equivalent to the schema used to build the Cloud Foundry
forms.

If a source platform is kind of "cloudfoundry", and the filter
schema pulled from hub is functionally equivalent to the schema
the components are based on, then show the custom components.  If
the schema is not equivalent, then show the generic schema defined
fields components.  This will most likely give the user a code editor
to input a json document to define the filter.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>