:bug: Fix custom target order handling
Resolves: https://issues.redhat.com/browse/MTA-6287
Resolves: #2708

All targets will new be displayed, even if they are not in the
target order setting.  The target query hook now will return the
server ordering of the targets, and ordering by the "ui.target.order"
setting.  All targets are included in both lists.  Any unordered
targets is added to the end of the ordered list.  By putting this
logic in the target query hook, target order can be used the same
way everywhere it is needed.

On the migration-targets page, the targets are now displayed in
the order specified by the "ui.target.order" setting.  Any targets
not in the setting are added to the end of the list.

On the custom-target-form, updated the onSubmit handler to
both create a new target and add it to the target order.  This
is done awaiting async mutations to help ensure the operations
finish before the create modal is closed.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>