[release-0.8] :bug: Update Cloud Foundry discover applications filter (#2718) (#2722)
Resolves: https://issues.redhat.com/browse/MTA-6326
Resolves: #2709
Backport-Of: #2718
CloudFoundry schema update:
https://github.com/konveyor/operator/pull/484

## Summary

Update the discover application wizard CloudFoundry source platform
filter input and review to support organizations, spaces, and names.

Added the ability to validate that two JSON schemas are functionally
equivalent. This is used to validate that the Cloud Foundry schema is
functionally equivalent to the schema used to build the Cloud Foundry
forms.

If a source platform is kind of "cloudfoundry", and the filter schema
pulled from hub is functionally equivalent to the schema the components
are based on, then show the custom components. If the schema is not
equivalent, then show the generic schema defined fields components. This
will most likely give the user a code editor to input a json document to
define the filter.

## Summary by CodeRabbit

* **New Features**
  * Added Cloud Foundry discovery filters: separate Organizations, Spaces,
    and Names in import and review flows.

* **Chores**
  * Updated dependency declarations and added lodash-es integration.
  * Moved and consolidated JSON Schema utilities for clearer schema
    handling and validation.

* **Tests**
  * Expanded JSON Schema utility test coverage and removed tests for a
    deprecated combine helper.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>