:test_tube: Update workflow to run cypress tests for PRs/pushes
This workflow allows PRs to specify test tags to use for the E2E tests
in the PR body.  By adding a line like this to the PR body:

```
ci test tags: @ci,@tier0,@tier1
```

the workflow will run the cypress tests for the ci, tier0 and tier1
tagged tests for this PR.

When the workflow is run outside of a PR, or the PR body does not
have the string, the workflow will use the default tags of
`@ci,@tier0`.

Part of: #2670

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>