:test_tube: Update cypress version to 15.7.0
The version of cypress was installed as 15.7.0 in the cypress package.json
file, but the package-lock.json file had two different versions of cypress
depending on the directory. Updated the package-lock.json to have a single
version, and updated the `cypress-fs` dependency to version 0.2.8 to
allow npm to install the correct version of cypress.

Running the tests locally from the cypress directory was running
version v15, but running the tests locally from the root directory was
running version v14.  This seems to be the root of the CI problem
encountered.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>