:test_tube: Update cypress to 15.7.0, use sourcemaps, add test spec handling (#2787)
The version of cypress was installed as 15.7.0 in the cypress
package.json file, but the package-lock.json file had two different
versions of cypress depending on the directory. Updated the
package-lock.json to have a single version, and updated the `cypress-fs`
dependency to version 0.2.8 to allow npm to install the correct version
of cypress.

Running the tests locally from the cypress directory was running version
v15, but running the tests locally from the root directory was running
version v14. This seems to be the root of the CI problem encountered. CI
testing uses the version of cypress available to the root of the
project.

## Summary by CodeRabbit

* **Chores**
  * Bumped a dev dependency to a newer patch release.
  * Enhanced CI to compute and wire the appropriate UI-test ref into the
    global workflow.

* **New Features**
  * Added a step to build test specs from provided test tags for UI test
    runs.
  * Normalized tier/tag inputs so tag formatting no longer affects test
    selection.
  * Sanitized test-ref names used for artifacts to avoid invalid
    characters.

* **Bug Fixes / Improvements**
  * Enabled sourcemaps and source content in test bundling for better
    debugging.
  * Improved logging and consolidated error reporting around test data
    seeding.

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>