:test_tube: Fix analysis wizard test flake with next button
WHen running the an analysis wizard test, the next button
on the custom rules step was not immediately enabled.  The form
requires a few renders to properly validate and enable the
Next button.  The tests would click right away, and sometimes
it would be enabled and sometimes not.  This would cause the
next "Next" click to hit the custom rules step instead of the
advanced options step.  In that case, the test would fail.

This fix adds a wait for the next button to be enabled before
clicking it.  This wait allows the form to validate, update state,
and the next buttons to be enabled.

First noticed in CI working on #2747

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>